move sdcCar high level logic to sdcHLC.cc
