+++ sdcIntersection.hh
@@ -4,14 +4,11 @@
 #include <vector>
 #include <iostream>
 #include <limits>
-#include "sdcWaypoint.hh"
 
-enum Direction {
-  north,
-  south,
-  east,
-  west
-};
+#include "sdcAngle.hh"
+#include "sdcWaypoint.hh"
+#include "constants.hh"
+
 enum WaypointType {
     WaypointType_DriveStraight,
     WaypointType_TurnLeft,
@@ -20,16 +17,16 @@ enum WaypointType {
 };
 
 class sdcIntersection {
-  public:
-  // neighbor = pair.first
-  // dist = pair.second
-  std::vector<std::pair<int, double>> neighbors_pairs;
-  int place;
-  int previous;
-  double dist = std::numeric_limits<double>::infinity();
-  bool visited = 0;
-  WaypointType wpType;
-  sdcWaypoint waypoint;
+    public:
+    // neighbor = pair.first
+    // dist = pair.second
+    std::vector<std::pair<int, double>> neighbors_pairs;
+    int place;
+    int previous;
+    double dist = std::numeric_limits<double>::infinity();
+    bool visited = 0;
+    WaypointType wpType;
+    sdcWaypoint waypoint;
 };
 
 #endif
